requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
defusedxml==0.7.1
Django==3.2.9
django-cors-headers==3.10.1
django-filter==21.1
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
httpie==2.6.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
python-magic==0.4.24
python-magic-bin==0.4.14
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
stripe==2.63.0
urllib3==1.26.7
zipp==3.6.0
